requirements.txt
alembic==1.4.2
appdirs==1.4.3
Babel==2.8.0
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
dominate==2.5.1
elasticsearch==7.8.1
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.18
urllib3==1.25.8
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.1
